requirements.txt
asgiref==3.6.0
bleach==6.0.0
Django==4.1.5
djangorestframework==3.14.0
drf-api-logger==1.1.12
python-decouple==3.6
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
webencodings==0.5.1
